Refactor how companion states are created (#309)
Fixes #307 

This PR changes how the companion states are created for ComposeTraits that have states (input value) such as TextField, Slider, Switch, etc.

Before this change, the companion state was created when the ComposeNode was dropped to the UiBuilder, but its implication often introduced the LLM-generated screens, not follow the implication resulting in build errors.

This PR changes the behavior to remove the implicit operation to create the companion state. Instead, specific ComposeTraits or Actions have companion states by default.