Introduce ValueFromCompanionState for AssignableProperty that depends on (#311)
* Introduce ValueFromCompanionState for AssignableProperty that depends on

the default companion state.

To reduce the implicit dependency to a state even though it's a default
companion state, the AssignableProperty needed to specify the ID of the
state, which LLM-generated screen often mis-generates.

This change is to reduce the complexity around the state dependnecy when
it depndson the default companion state

* Add missing function call